se agrego un circulo en la memoria virtual y una memoria de constantes y que se pueda leer el archivo de Objecto ejecutable

+++ compile/Semantica.h
@@ -187,7 +187,7 @@ public :
 	int para_flo;
 	int para_bool;
 	Memory();
-	Direction* save(int partition, int type, float value);
+	Direction* save(int partition, int type, double value);
 	void clearLocal();
 	void clearTemp();
 	void clearPara();
